+++ src/index.js
@@ -11,13 +11,16 @@ import { resolve as resolve_url } from 'url';
 import http from 'http';
 import https from 'https';
 import zlib from 'zlib';
-import { PassThrough } from 'stream';
+import Stream from 'stream';
 
 import Body, { writeToStream, getTotalBytes } from './body';
 import Response from './response';
 import Headers, { createHeadersLenient } from './headers';
 import Request, { getNodeRequestOptions } from './request';
 import FetchError from './fetch-error';
+
+// fix an issue where PassThrough isn't a named export for node <10
+const PassThrough = Stream.PassThrough;
 
 /**
  * Fetch function